backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
pydantic[email]
psycopg2

